pyproject.toml
[project]
name = "app"
version = "0.1.0"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "anthropic>=0.67.0",
    "duckdb>=1.3.2",
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "langgraph>=0.5.1",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langsmith>=0.4.4",
    "loguru>=0.7.3",
    "mcp[cli]>=1.2.0",
    "nanoid>=2.0.0",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic-settings>=2.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length=79

[tool.ruff.lint]
extend-select = ["I"]

[tool.mypy]
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pyright]
reportMissingTypeStubs = false

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "tqdm>=4.67.1",
]
lint = [
    "mypy>=1.17.1",
    "ruff>=0.12.8",
]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
]
